Add files via upload
This is the bricks functionality class, which aims to create the bricks for the brick breaker game. There are different bricks of different colors, indicating their hardness means after how many collisions it has to be destroyed. Moreover for convenience to integrate with the bricks a sample ball class is also present in this class.
Note: "For the programmer who is integrating the ball class with the brick class, it is advisable to see how the ball randomly goes through movement patterns, and the progarmmer who is controlling the flow control, look for the images bath to adjust it accordingly".